Updated oembed service
