+++ packages/server-core/src/media/oembed/oembed.ts
@@ -23,13 +23,27 @@ All portions of the code written by the Ethereal Engine team are Copyright © 20
 Ethereal Engine. All Rights Reserved.
 */
 
-import { archiverQuerySchema } from '@etherealengine/engine/src/schemas/media/archiver.schema'
-import { createSwaggerServiceOptions } from 'feathers-swagger'
+import { oembedMethods, oembedPath } from '@etherealengine/engine/src/schemas/media/oembed.schema'
+import { Application } from '../../../declarations'
+import { OembedService } from './oembed.class'
+import oembedDocs from './oembed.docs'
+import hooks from './oembed.hooks'
 
-export default createSwaggerServiceOptions({
-  schemas: { archiverQuerySchema },
-  docs: {
-    description: 'Archiver service description',
-    securities: ['all']
+declare module '@etherealengine/common/declarations' {
+  interface ServiceTypes {
+    [oembedPath]: OembedService
   }
-})
+}
+
+export default (app: Application): void => {
+  app.use(oembedPath, new OembedService(app), {
+    // A list of all methods this service exposes externally
+    methods: oembedMethods,
+    // You can add additional custom events to be sent to clients here
+    events: [],
+    docs: oembedDocs
+  })
+
+  const service = app.service(oembedPath)
+  service.hooks(hooks)
+}
