+++ packages/server-core/src/media/services.ts
@@ -25,7 +25,7 @@ Ethereal Engine. All Rights Reserved.
 
 import FileBrowserUpload from './file-browser-upload/file-browser-upload'
 import FileBrowser from './file-browser/file-browser'
-import OEmbed from './oembed/oembed.service'
+import OEmbed from './oembed/oembed'
 import Archiver from './recursive-archiver/archiver'
 import StaticResourceFilters from './static-resource-filters/static-resource-filters'
 import StaticResource from './static-resource/static-resource'
